Remove std dependency from wasmlanche-sdk
