+++ x/programs/rust/wasmlanche-sdk/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 borsh = { version = "1.5.1", features = ["derive"] }
 bytemuck = { version = "1.17.0", features = ["derive"] }
+displaydoc = { version = "0.2.5", default-features = false }
 hashbrown = "0.14.5"
 sdk-macros = { workspace = true }
 thiserror = { workspace = true }
