+++ x/programs/rust/wasmlanche-sdk/src/state.rs
@@ -12,21 +12,20 @@ use alloc::{boxed::Box, vec::Vec};
 use borsh::{from_slice, BorshDeserialize, BorshSerialize};
 use bytemuck::NoUninit;
 use core::mem::{self, size_of};
+use displaydoc::Display;
 use hashbrown::HashMap;
 use sdk_macros::impl_to_pairs;
 
 // maximum number of chunks that can be stored at the key as big endian u16
 pub const STATE_MAX_CHUNKS: [u8; 2] = 4u16.to_be_bytes();
 
-#[derive(Clone, thiserror::Error, Debug)]
+#[derive(Clone, Debug, Display)]
 pub enum Error {
-    #[error("invalid byte length: {0}")]
+    /// invalid byte length {0}
     InvalidByteLength(usize),
-
-    #[error("failed to serialize bytes")]
+    /// failed to serialize bytes
     Serialization,
-
-    #[error("failed to deserialize bytes")]
+    /// failed to deserialize bytes
     Deserialization,
 }
